Introduce validation annotations (ValidPassword, ValidDateRange) and refactor password and date validation logic. Simplify service and controller methods by removing redundant code.

+++ backend/src/main/java/com/ajaros/reservationsystem/equipment/services/EquipmentService.java
@@ -40,10 +40,13 @@ public class EquipmentService {
   public EquipmentResponse updateEquipment(Long id, String newName) {
     var equipment = getEquipmentById(id);
     equipment.setName(newName);
-    return equipmentMapper.toEquipmentResponse(equipmentRepository.save(equipment));
+    return equipmentMapper.toEquipmentResponse(equipment);
   }
 
   public void deleteEquipment(Long id) {
+    if (!equipmentRepository.existsById(id)) {
+      throw new EquipmentNotFoundException(id);
+    }
     equipmentRepository.deleteById(id);
   }
 }
